+++ .go-coverage-baseline.json
@@ -1,9 +1,9 @@
 {
-  "minTotalCoverage": 44.6,
+  "minTotalCoverage": 44.8,
   "minCoverageByPackage": {
     "pkg/authentication": 6.3,
     "pkg/authorization": 52.4,
-    "pkg/cli": 27.8,
+    "pkg/cli": 44.4,
     "pkg/cli/commands/canvases": 35.9,
     "pkg/cli/commands/canvases/models": 100,
     "pkg/cli/commands/events": 39.1,
@@ -29,13 +29,13 @@
     "pkg/grpc/actions/agents": 44,
     "pkg/grpc/actions/auth": 80.8,
     "pkg/grpc/actions/blueprints": 0,
-    "pkg/grpc/actions/canvases": 62.1,
+    "pkg/grpc/actions/canvases": 63,
     "pkg/grpc/actions/canvases/changesets": 84.7,
-    "pkg/grpc/actions/canvases/layout": 63.4,
+    "pkg/grpc/actions/canvases/layout": 63.7,
     "pkg/grpc/actions/integrations": 0,
     "pkg/grpc/actions/me": 54.2,
     "pkg/grpc/actions/messages": 0,
-    "pkg/grpc/actions/organizations": 67.2,
+    "pkg/grpc/actions/organizations": 67,
     "pkg/grpc/actions/secrets": 42.6,
     "pkg/grpc/actions/serviceaccounts": 5.5,
     "pkg/grpc/actions/triggers": 84.6,
@@ -44,13 +44,13 @@
     "pkg/jwt": 55.3,
     "pkg/lint/examplepayloads": 84.6,
     "pkg/logging": 0,
-    "pkg/models": 9.8,
-    "pkg/networkpolicy": 82.1,
+    "pkg/models": 9.1,
+    "pkg/networkpolicy": 87.9,
     "pkg/oidc": 70.4,
-    "pkg/public": 51.9,
+    "pkg/public": 51.8,
     "pkg/public/middleware": 46.4,
     "pkg/public/ws": 0,
-    "pkg/registry": 82.6,
+    "pkg/registry": 82.9,
     "pkg/retry": 0,
     "pkg/secrets": 62.5,
     "pkg/server": 0,
@@ -62,7 +62,7 @@
     "pkg/web": 82.1,
     "pkg/widgets/annotation": 0,
     "pkg/workers": 41,
-    "pkg/workers/contexts": 46.4,
+    "pkg/workers/contexts": 45.9,
     "pkg/workers/eventdistributer": 0
   },
   "ignoredPackagePrefixes": [
@@ -73,5 +73,5 @@
     "pkg/triggers",
     "pkg/web/assets"
   ],
-  "updatedAt": "2026-05-03T23:40:42Z"
+  "updatedAt": "2026-05-06T17:45:29Z"
 }
