+++ pkg/authorization/interceptor.go
@@ -281,11 +281,6 @@ func NewAuthorizationInterceptor(authService Authorization) *AuthorizationInterc
 			Action:     "update",
 			DomainType: models.DomainTypeOrganization,
 		},
-		pbCanvases.Canvases_UpdateCanvasFolderPosition_FullMethodName: {
-			Resource:   "canvases",
-			Action:     "update",
-			DomainType: models.DomainTypeOrganization,
-		},
 		pbCanvases.Canvases_UpdateCanvasFolderMembership_FullMethodName: {
 			Resource:         "canvases",
 			Action:           "update",
