+++ pkg/grpc/actions/canvases/update_canvas_folder_test.go
@@ -25,12 +25,18 @@ func Test__UpdateCanvasFolder__UpdatesFolder(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	updateResponse, err := UpdateCanvasFolder(ctx, r.Organization.ID.String(), createResponse.Folder.Metadata.Id, &pb.CanvasFolder{
-		Spec: &pb.CanvasFolder_Spec{
-			Title:           "Production Ops",
-			BackgroundColor: models.CanvasFolderColor3,
+	updateResponse, err := UpdateCanvasFolder(
+		ctx,
+		r.Organization.ID.String(),
+		createResponse.Folder.Metadata.Id,
+		&pb.CanvasFolder{
+			Spec: &pb.CanvasFolder_Spec{
+				Title:           "Production Ops",
+				BackgroundColor: models.CanvasFolderColor3,
+			},
 		},
-	})
+		pb.UpdateCanvasFolderRequest_DIRECTION_UNSPECIFIED,
+	)
 	require.NoError(t, err)
 	require.NotNil(t, updateResponse.Folder)
 	require.NotNil(t, updateResponse.Folder.Spec)
@@ -52,12 +58,96 @@ func Test__UpdateCanvasFolder__RejectsDuplicateTitle(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	_, err = UpdateCanvasFolder(ctx, r.Organization.ID.String(), secondFolder.Folder.Metadata.Id, &pb.CanvasFolder{
-		Spec: &pb.CanvasFolder_Spec{
-			Title:           firstFolder.Folder.Spec.Title,
-			BackgroundColor: models.CanvasFolderColor1,
+	_, err = UpdateCanvasFolder(
+		ctx,
+		r.Organization.ID.String(),
+		secondFolder.Folder.Metadata.Id,
+		&pb.CanvasFolder{
+			Spec: &pb.CanvasFolder_Spec{
+				Title:           firstFolder.Folder.Spec.Title,
+				BackgroundColor: models.CanvasFolderColor1,
+			},
 		},
-	})
+		pb.UpdateCanvasFolderRequest_DIRECTION_UNSPECIFIED,
+	)
 	require.Error(t, err)
 	assert.Equal(t, codes.AlreadyExists, status.Code(err))
+}
+
+func Test__UpdateCanvasFolder__MovesFolderUpAndDown(t *testing.T) {
+	r := support.Setup(t)
+	ctx := context.Background()
+
+	firstFolder, err := CreateCanvasFolder(ctx, r.Organization.ID.String(), &pb.CanvasFolder{
+		Spec: &pb.CanvasFolder_Spec{Title: "First"},
+	})
+	require.NoError(t, err)
+
+	secondFolder, err := CreateCanvasFolder(ctx, r.Organization.ID.String(), &pb.CanvasFolder{
+		Spec: &pb.CanvasFolder_Spec{Title: "Second"},
+	})
+	require.NoError(t, err)
+
+	thirdFolder, err := CreateCanvasFolder(ctx, r.Organization.ID.String(), &pb.CanvasFolder{
+		Spec: &pb.CanvasFolder_Spec{Title: "Third"},
+	})
+	require.NoError(t, err)
+
+	moveUpResponse, err := UpdateCanvasFolder(
+		ctx,
+		r.Organization.ID.String(),
+		secondFolder.Folder.Metadata.Id,
+		nil,
+		pb.UpdateCanvasFolderRequest_DIRECTION_UP,
+	)
+	require.NoError(t, err)
+	require.Len(t, moveUpResponse.Folders, 3)
+	assert.Equal(t, []string{
+		secondFolder.Folder.Metadata.Id,
+		thirdFolder.Folder.Metadata.Id,
+		firstFolder.Folder.Metadata.Id,
+	}, []string{
+		moveUpResponse.Folders[0].Metadata.Id,
+		moveUpResponse.Folders[1].Metadata.Id,
+		moveUpResponse.Folders[2].Metadata.Id,
+	})
+
+	moveDownResponse, err := UpdateCanvasFolder(
+		ctx,
+		r.Organization.ID.String(),
+		secondFolder.Folder.Metadata.Id,
+		nil,
+		pb.UpdateCanvasFolderRequest_DIRECTION_DOWN,
+	)
+	require.NoError(t, err)
+	require.Len(t, moveDownResponse.Folders, 3)
+	assert.Equal(t, []string{
+		thirdFolder.Folder.Metadata.Id,
+		secondFolder.Folder.Metadata.Id,
+		firstFolder.Folder.Metadata.Id,
+	}, []string{
+		moveDownResponse.Folders[0].Metadata.Id,
+		moveDownResponse.Folders[1].Metadata.Id,
+		moveDownResponse.Folders[2].Metadata.Id,
+	})
+}
+
+func Test__UpdateCanvasFolder__RejectsMixedFieldAndPositionUpdates(t *testing.T) {
+	r := support.Setup(t)
+	ctx := context.Background()
+
+	folder, err := CreateCanvasFolder(ctx, r.Organization.ID.String(), &pb.CanvasFolder{
+		Spec: &pb.CanvasFolder_Spec{Title: "Production"},
+	})
+	require.NoError(t, err)
+
+	_, err = UpdateCanvasFolder(
+		ctx,
+		r.Organization.ID.String(),
+		folder.Folder.Metadata.Id,
+		&pb.CanvasFolder{Spec: &pb.CanvasFolder_Spec{Title: "Production Ops"}},
+		pb.UpdateCanvasFolderRequest_DIRECTION_UP,
+	)
+	require.Error(t, err)
+	assert.Equal(t, codes.InvalidArgument, status.Code(err))
 }
