+++ web_src/src/pages/home/CanvasCardsGrid.tsx
@@ -1,0 +1,205 @@
+import { Heading } from "@/components/Heading/heading";
+import { Text } from "@/components/Text/text";
+import type { SuperplaneComponentsEdge, SuperplaneComponentsNode } from "@/api-client";
+import { Link } from "react-router-dom";
+import { CanvasActionsMenu } from "./CanvasActionsMenu";
+import type { CanvasCardData, CanvasFolderData } from "./types";
+
+interface CanvasCardsGridProps {
+  canvases: CanvasCardData[];
+  canvasFolders: CanvasFolderData[];
+  organizationId: string;
+  onEditCanvas: (canvas: CanvasCardData) => void;
+  canUpdateCanvases: boolean;
+  canDeleteCanvases: boolean;
+  permissionsLoading: boolean;
+}
+
+export function CanvasCardsGrid({
+  canvases,
+  canvasFolders,
+  organizationId,
+  onEditCanvas,
+  canUpdateCanvases,
+  canDeleteCanvases,
+  permissionsLoading,
+}: CanvasCardsGridProps) {
+  return (
+    <div className="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-4">
+      {canvases.map((canvas) => (
+        <CanvasCard
+          key={canvas.id}
+          canvas={canvas}
+          canvasFolders={canvasFolders}
+          organizationId={organizationId}
+          onEdit={onEditCanvas}
+          canUpdateCanvases={canUpdateCanvases}
+          canDeleteCanvases={canDeleteCanvases}
+          permissionsLoading={permissionsLoading}
+        />
+      ))}
+    </div>
+  );
+}
+
+interface CanvasCardProps {
+  canvas: CanvasCardData;
+  canvasFolders: CanvasFolderData[];
+  organizationId: string;
+  onEdit: (canvas: CanvasCardData) => void;
+  canUpdateCanvases: boolean;
+  canDeleteCanvases: boolean;
+  permissionsLoading: boolean;
+}
+
+function CanvasCard({
+  canvas,
+  canvasFolders,
+  organizationId,
+  onEdit,
+  canUpdateCanvases,
+  canDeleteCanvases,
+  permissionsLoading,
+}: CanvasCardProps) {
+  const canvasHref = `/${organizationId}/canvases/${canvas.id}`;
+  const previewNodes = canvas.nodes || [];
+  const previewEdges = canvas.edges || [];
+
+  return (
+    <div className="relative min-h-40 bg-white dark:bg-gray-800 rounded-md outline outline-gray-950/15 hover:shadow-md transition-shadow cursor-pointer">
+      <Link to={canvasHref} aria-label={`Open canvas ${canvas.name}`} className="absolute inset-0 rounded-md" />
+      <div className="pointer-events-none relative flex flex-col h-full">
+        <div className="p-3">
+          <div className="flex items-start justify-between gap-3">
+            <div className="flex flex-col flex-1 min-w-0">
+              <Heading
+                level={3}
+                className="mb-0 line-clamp-2 !text-sm font-medium text-gray-800 transition-colors !leading-5"
+              >
+                <span className="truncate">{canvas.name}</span>
+              </Heading>
+            </div>
+            <div className="pointer-events-auto">
+              <CanvasActionsMenu
+                canvas={canvas}
+                canvasFolders={canvasFolders}
+                organizationId={organizationId}
+                onEdit={onEdit}
+                canUpdateCanvases={canUpdateCanvases}
+                canDeleteCanvases={canDeleteCanvases}
+                permissionsLoading={permissionsLoading}
+              />
+            </div>
+          </div>
+
+          {canvas.description ? (
+            <div className="mb-3">
+              <Text className="line-clamp-2 text-left text-[12px] !leading-normal text-gray-800 dark:text-gray-400">
+                {canvas.description}
+              </Text>
+            </div>
+          ) : null}
+
+          <div className="flex justify-between items-center">
+            <p className="mt-1 text-left text-[11px] leading-none text-gray-500 dark:text-gray-400">
+              {canvas.createdBy?.name ? (
+                <>
+                  Created by {canvas.createdBy.name}, on {canvas.createdAt}
+                </>
+              ) : (
+                <>Created on {canvas.createdAt}</>
+              )}
+            </p>
+          </div>
+        </div>
+
+        <CanvasMiniMap nodes={previewNodes} edges={previewEdges} />
+      </div>
+    </div>
+  );
+}
+
+interface CanvasMiniMapProps {
+  nodes?: SuperplaneComponentsNode[];
+  edges?: SuperplaneComponentsEdge[];
+}
+
+function CanvasMiniMap({ nodes = [], edges = [] }: CanvasMiniMapProps) {
+  const positionedNodes = nodes.filter(
+    (node) => typeof node.position?.x === "number" && typeof node.position?.y === "number",
+  ) as Array<SuperplaneComponentsNode & { position: { x: number; y: number } }>;
+
+  if (!positionedNodes.length) {
+    return <div className="h-24 w-full p-3 pt-0" />;
+  }
+
+  const xs = positionedNodes.map((node) => node.position.x);
+  const ys = positionedNodes.map((node) => node.position.y);
+  const minX = Math.min(...xs);
+  const maxX = Math.max(...xs);
+  const minY = Math.min(...ys);
+  const maxY = Math.max(...ys);
+  const padding = 80;
+  const width = Math.max(maxX - minX, 200) + padding * 2;
+  const height = Math.max(maxY - minY, 200) + padding * 2;
+  const viewBox = `${minX - padding} ${minY - padding} ${width} ${height}`;
+  const nodeWidth = Math.min(Math.max(width * 0.08, 30), 80);
+  const nodeHeight = nodeWidth * 0.45;
+
+  const nodePositions = new Map<string, { x: number; y: number }>();
+  positionedNodes.forEach((node) => {
+    const id = node.id || node.name;
+    if (!id) return;
+    nodePositions.set(id, { x: node.position.x, y: node.position.y });
+  });
+
+  const drawableEdges =
+    edges?.filter(
+      (edge) => edge.sourceId && edge.targetId && nodePositions.has(edge.sourceId) && nodePositions.has(edge.targetId),
+    ) || [];
+
+  return (
+    <div className="w-full overflow-hidden p-3 pt-0">
+      <svg
+        viewBox={viewBox}
+        preserveAspectRatio="xMidYMid meet"
+        className="h-24 w-full text-gray-500 dark:text-gray-400"
+      >
+        {drawableEdges.map((edge) => {
+          const source = nodePositions.get(edge.sourceId!);
+          const target = nodePositions.get(edge.targetId!);
+          if (!source || !target) return null;
+          return (
+            <line
+              key={`${edge.sourceId}-${edge.targetId}`}
+              x1={source.x}
+              y1={source.y}
+              x2={target.x}
+              y2={target.y}
+              stroke="currentColor"
+              strokeWidth={6}
+              strokeLinecap="round"
+              opacity={0.25}
+            />
+          );
+        })}
+        {positionedNodes.map((node) => {
+          const id = node.id || node.name || `${node.position.x}-${node.position.y}`;
+          return (
+            <rect
+              key={id}
+              x={node.position.x - nodeWidth / 2}
+              y={node.position.y - nodeHeight / 2}
+              width={nodeWidth}
+              height={nodeHeight}
+              rx={8}
+              ry={8}
+              fill="#1f2937"
+              opacity={1}
+            />
+          );
+        })}
+      </svg>
+    </div>
+  );
+}
